requirements.txt
fastapi
uvicorn
pymysql
aiomysql
pydantic
sqlalchemy
python-dotenv
alembic
cryptography
pytest
httpx
pytest-mock
cachetools
